Refactor: ErrorResponse를 ApiResponse로 감싸서 응답
