+++ src/main/java/com/kdt/api/ApiExceptionAdvice.java
@@ -1,4 +1,4 @@
-package com.kdt.aop;
+package com.kdt.api;
 
 import static org.springframework.http.HttpStatus.BAD_REQUEST;
 import static org.springframework.http.HttpStatus.NOT_FOUND;
@@ -6,7 +6,6 @@ import static org.springframework.http.HttpStatus.NOT_FOUND;
 import com.kdt.exception.NotFoundException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
-import org.springframework.validation.BindException;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
@@ -18,19 +17,19 @@ public class ApiExceptionAdvice {
     @ExceptionHandler(Exception.class)
     public ResponseEntity globalException(Exception e) {
         log.error("Unexpected Exception : {}", e.getMessage());
-        return ResponseEntity.badRequest().body(ErrorResponse.of(e.getMessage()));
+        return ResponseEntity.badRequest().body(ApiResponse.of(ErrorResponse.of(e.getMessage())));
     }
 
     @ExceptionHandler(NotFoundException.class)
     public ResponseEntity badRequestException(Exception e) {
         log.error("Unexpected Exception : {}", e.getMessage());
-        return ResponseEntity.status(NOT_FOUND).body(ErrorResponse.of(e.getMessage()));
+        return ResponseEntity.status(NOT_FOUND).body(ApiResponse.of(ErrorResponse.of(e.getMessage())));
     }
 
     @ExceptionHandler(MethodArgumentNotValidException.class)
     public ResponseEntity bindException(MethodArgumentNotValidException e) {
         log.error("Unexpected Exception : {}", e.getMessage());
-        return ResponseEntity.status(BAD_REQUEST).body(ErrorResponse.of(e.getBindingResult()));
+        return ResponseEntity.status(BAD_REQUEST).body(ApiResponse.of(ErrorResponse.of(e.getBindingResult())));
     }
 
 }
