+++ src/main/java/com/kdt/api/PostApi.java
@@ -1,18 +1,15 @@
 package com.kdt.api;
 
 import static com.kdt.api.PostApi.POSTS;
-import static org.springframework.http.HttpStatus.BAD_REQUEST;
 import static org.springframework.http.HttpStatus.CREATED;
 import static org.springframework.http.HttpStatus.NO_CONTENT;
 
-import com.kdt.aop.ErrorResponse;
 import com.kdt.post.dto.PostDto;
 import com.kdt.post.service.PostService;
 import javax.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
-import org.springframework.validation.Errors;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
